StrandsAgent/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
python-dotenv>=1.0.1
psycopg[binary]>=3.1.18
psycopg-pool>=3.2.0
